fix: mocked type error
