+++ __tests__/components/sections/challenges/Header.test.tsx
@@ -2,8 +2,7 @@ import ChallengeHeader from "@/components/sections/challenges/Header";
 import "@testing-library/jest-dom";
 import { screen } from "@testing-library/react";
 import { renderWithRedux } from "../../../../__mocks__/renderWithRedux";
-import { challenge } from "../../../../__mocks__/challenge";
-import { submission } from "../../../../__mocks__/submission";
+import { challenge, submission } from "../../../../__mocks__/challenge";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -18,11 +17,12 @@ jest.mock("next/router", () => ({
 
 describe("ChallengeHeader", () => {
   it("should render the challenge header with section", () => {
-    renderWithRedux(<ChallengeHeader />, { challenges: { current: challenge, list: [challenge], loading: false, submission: submission } });
+    renderWithRedux(<ChallengeHeader />, { challenges: { current: challenge(), list: [challenge()], loading: false, submission: submission() } });
+    const challenges = challenge();
     const challengeheader = screen.getByTestId("challengeHeaderId");
     expect(challengeheader).toBeInTheDocument();
     expect(screen.getByText("communities.challenge.title")).toBeInTheDocument();
-    expect(screen.getByText(challenge.name)).toBeInTheDocument();
-    expect(screen.getByText(challenge.description)).toBeInTheDocument();
+    expect(screen.getByText(challenges.name)).toBeInTheDocument();
+    expect(screen.getByText(challenges.description)).toBeInTheDocument();
   });
 });
