+++ __tests__/components/sections/challenges/Learning.test.tsx
@@ -6,8 +6,7 @@ import { renderWithRedux } from "../../../../__mocks__/renderWithRedux";
 import {  mockCourse, mockLearningModule } from "../../../../__mocks__/course";
 import { setupServer } from "msw/node";
 import { handlers } from "../../../../__mocks__/provider/handlers";
-import { submission } from "../../../../__mocks__/submission";
-import { challenge } from "../../../../__mocks__/challenge";
+import { challenge, submission } from "../../../../__mocks__/challenge";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -37,7 +36,7 @@ describe("Learning", () => {
   it("should render with courses", () => {
     const coursesArray = [mockCourse];
     renderWithRedux(<Learning courses={[mockCourse]} learningModules={[mockLearningModule]} community={community} />, {
-      challenges: { current: challenge, list: [challenge], loading: false, submission: submission },
+      challenges: { current: challenge(), list: [challenge()], loading: false, submission: submission() },
     });
     coursesArray.forEach((course) => {
       expect(screen.getByText(course.name)).toBeInTheDocument();
