+++ __tests__/components/sections/challenges/SetUpTeamChallenge.test.tsx
@@ -2,8 +2,7 @@ import "@testing-library/jest-dom";
 import { screen } from "@testing-library/react";
 import { renderWithRedux } from "../../../../__mocks__/renderWithRedux";
 import SetupTeamChallenge from "@/components/sections/challenges/SetupTeamChallenge";
-import { challenge, mockInvite } from "../../../../__mocks__/challenge";
-import { submission } from "../../../../__mocks__/submission";
+import { challenge,submission, mockInvite } from "../../../../__mocks__/challenge";
 
 jest.mock("next/router", () => ({
   useRouter: () => ({
@@ -22,20 +21,21 @@ describe("SetupTeamChallenge", () => {
   });
 
   it("renders content correctly", () => {
-    renderWithRedux(<SetupTeamChallenge />, { challenges: { current: challenge, list: [challenge], loading: false, submission: submission }, invites: { data: mockInvite } });
+    renderWithRedux(<SetupTeamChallenge />, { challenges: { current: challenge(), list: [challenge()], loading: false, submission: submission() }, invites: { data: mockInvite } });
     expect(screen.getByText("Submission")).toBeInTheDocument();
     expect(screen.getByText("communities.overview.challenge.team.setup.info")).toBeInTheDocument();
     expect(screen.getByText("Form your team")).toBeInTheDocument();
   });
 
   it("renders CreateTeamCard when there is no invitation or comfirmTeamInvitation when there is invitation", () => {
-    renderWithRedux(<SetupTeamChallenge />, { challenges: { current: challenge, list: [challenge], loading: false, submission: submission }, invites: { data: mockInvite } });
+    renderWithRedux(<SetupTeamChallenge />, { challenges: { current: challenge(), list: [challenge()], loading: false, submission: submission() }, invites: { data: mockInvite } });
+    const challenges = challenge();
     if (!mockInvite) {
       expect(screen.getByText("communities.overview.challenge.team.setup.submit-title" || "")).toBeInTheDocument();
       expect(screen.getByText("communities.overview.challenge.team.setup.description" || "")).toBeInTheDocument();
     } else {
       expect(screen.getByText("Submit your team")).toBeInTheDocument();
-      expect(screen.getByText(`The maximum team members for this challenge is ${challenge?.teamLimit || "3"} people`)).toBeInTheDocument();
+      expect(screen.getByText(`The maximum team members for this challenge is ${challenges?.teamLimit || "3"} people`)).toBeInTheDocument();
     }
   });
 });
